few changes to the about structure
