+++ assets/static_data/choices.json
@@ -1,9 +1,5 @@
 {
-<<<<<<< HEAD
-  "name": "Choices",
-=======
   "name": "choice",
->>>>>>> 3316c18e00215b983a499d298c7df846a9644782
   "choices": [
     {
       "title": "OECD",
@@ -133,9 +129,5 @@
       ]
     }
   ]
-<<<<<<< HEAD
-}
-=======
 
 }
->>>>>>> 3316c18e00215b983a499d298c7df846a9644782
