+++ lib/routes.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
-
+import 'package:learning_compass_exp/screens/about/about.dart';
 import 'package:learning_compass_exp/screens/home/index.dart';
 import 'package:learning_compass_exp/screens/info/index.dart';
-
 import 'package:learning_compass_exp/data/models/petal_names.dart';
 
-import 'screens/home/widgets/drawer/tabbed_app_bar/tabbed_bar.dart';
 
 final routes = <String, WidgetBuilder>{
   '/': (context) => HomeScreen(),
@@ -44,5 +42,5 @@ final routes = <String, WidgetBuilder>{
   ),
 
   //Drawer
-  '/drawer/about': (context) => AboutTabbedAppBar(),
+  '/drawer/about': (context) => About(),
 };
