+++ test/widget_tests/flower_widget_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:redux/redux.dart';
-import 'package:flutter_redux/flutter_redux.dart';
 import 'dart:math';
 
 import 'package:learning_compass_exp/screens/home/widgets/flower/flower.dart';
